Deprecated : Entity/Contact, use Model/Contact
The Travis tests have been adjusted to use maintained version of
Symfony.
