+++ .travis.yml
@@ -24,14 +24,19 @@ matrix:
     - php: 5.6
       env: SYMFONY_VERSION=2.8.*
     - php: 7.0
+      env: SYMFONY_VERSION=2.8.*
     - php: 7.1
       env: SYMFONY_VERSION=2.8.*
     - php: 7.0
-    - php: 7.1
-      env: SYMFONY_VERSION=3.2.*
-    - php: 7.0
+      env: SYMFONY_VERSION=3.3.*
     - php: 7.1
       env: SYMFONY_VERSION=3.3.*
+    - php: 7.0
+      env: SYMFONY_VERSION=3.4.*
+    - php: 7.1
+      env: SYMFONY_VERSION=3.4.*
+    - php: 7.1
+      env: SYMFONY_VERSION=4.1.*
   allow_failures:
     - php: hhvm
