+++ Entity/Contact.php
@@ -3,127 +3,19 @@
 namespace TLH\ContactBundle\Entity;
 
 use Doctrine\ORM\Mapping as ORM;
-use Symfony\Component\Validator\Constraints as Assert;
-use Symfony\Component\Validator\ExecutionContext;
+use TLH\ContactBundle\Model\Contact as ModelContact;
 
 /**
  * TLH\ContactBundle\Entity\Contact
  *
+ * @deprecated Use TLH\ContactBundle\Model\Contact instead
  * @ORM\MappedSuperclass()
  */
-class Contact
+class Contact extends ModelContact
 {
-    /**
-     * @var bigint $id
-     *
-     * @ORM\Column(name="id", type="integer")
-     * @ORM\Id
-     * @ORM\GeneratedValue(strategy="AUTO")
-     */
-    protected $id;
-
-    /**
-     * @var email
-     *
-     * @ORM\Column(name="email", type="string", length=255, nullable=false)
-     */
-    protected $email;
-
-    /**
-     * @var message
-     *
-     * @ORM\Column(name="message", type="text", nullable=false)
-     */
-    protected $message;
-
-    /**
-     * @var datetime
-     *
-     * @ORM\Column(name="creation", type="datetime", nullable=false)
-     */
-    protected $creation;
-
     public function __construct()
     {
+        parent::__construct();
         $this->creation = new \DateTime("now");
-    }
-
-    /**
-     * Get id
-     *
-     * @return integer 
-     */
-    public function getId()
-    {
-        return $this->id;
-    }
-
-    /**
-     * Set email
-     *
-     * @param string $email
-     * @return Contact
-     */
-    public function setEmail($email)
-    {
-        $this->email = $email;
-    
-        return $this;
-    }
-
-    /**
-     * Get email
-     *
-     * @return string 
-     */
-    public function getEmail()
-    {
-        return $this->email;
-    }
-
-    /**
-     * Set message
-     *
-     * @param string $message
-     * @return Contact
-     */
-    public function setMessage($message)
-    {
-        $this->message = $message;
-    
-        return $this;
-    }
-
-    /**
-     * Get message
-     *
-     * @return string 
-     */
-    public function getMessage()
-    {
-        return $this->message;
-    }
-
-    /**
-     * Set creation
-     *
-     * @param string $creation
-     * @return Contact
-     */
-    public function setCreation($creation)
-    {
-        $this->creation = $creation;
-    
-        return $this;
-    }
-
-    /**
-     * Get creation
-     *
-     * @return string 
-     */
-    public function getCreation()
-    {
-        return $this->creation;
     }
 }
