+++ composer.json
@@ -19,8 +19,9 @@
         "symfony/expression-language": "~2.6|~3.0|~4.0",
         "symfony/class-loader": "*",
         "symfony/framework-bundle": "^2.7|^3.3|^4.0",
-        "doctrine/orm": "*",
+        "doctrine/orm": "^2.3",
         "symfony/form": "^2.7|^3.3|^4.0",
+        "symfony/templating": "^2.7|^3.3|^4.0",
         "symfony/validator": "^2.7|^3.3|^4.0",
         "swiftmailer/swiftmailer": "~4.2|~5.0"
     },
